Move schema sampling to separate module and cleanup vPartition creation code
